requirements.txt
requests
kafka-python 
pymongo

